Refactor (apps/pet-shop): Move `products-service.type.ts`
