+++ apps/pet-shop/src/app/integrator/data-access/backend-api/ps-products/ps-products-http.service.ts
@@ -6,7 +6,7 @@ import { map } from 'rxjs/operators'
 // # Internal modules
 import type { PSProduct } from '../../../../../architecture/entities/ps-products/ps-products.type'
 import { environment } from '../../../../../environments/environment'
-import type { PSProductsServiceReadList } from '../../../ui/ps-products/products-service.type'
+import type { PSProductsServiceReadList } from '../../ps-products/products-service.type'
 import { BackendAPIService } from '../backend-api/backend-api.service'
 import type { PSProductForBE } from './ps-products-for-be.type'
