+++ apps/pet-shop/src/app/integrator/data-access/ps-products/products-service.type.ts
